ingestion/requirements.txt
requests>=2.31
clickhouse-connect>=0.7
python-dotenv>=1.0
